backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
openai==1.54.3
aiofiles==24.1.0
python-json-logger==2.0.7
pydantic-settings==2.5.2

